pyproject.toml
[project]
name = "masksdkagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "mask-privacy[spacy]>=1.0.0",
    "python-dotenv>=1.2.2",
]
